use existing functions class
