+++ lib/classes/AutoCompleteHandler.ts
@@ -10,18 +10,18 @@ import applicationCommandOptionTypeReference from "../utilities/reference.js";
 import type AutoComplete from "./AutoComplete";
 import type Language from "./Language";
 
-export default class AutoCompleteHandler {
+export default class AutoCompleteHandler<C extends ExtendedClient> {
 	/**
 	 * Our extended client.
 	 */
-	public readonly client: ExtendedClient;
+	public readonly client: C;
 
 	/**
 	 * Create our auto complete handler.
 	 *
 	 * @param client Our extended client.
 	 */
-	public constructor(client: ExtendedClient) {
+	public constructor(client: C) {
 		this.client = client;
 	}
