+++ lib/typings/options.ts
@@ -1,0 +1,6 @@
+import type ExtendedClient from "../extensions/ExtendedClient";
+import type Functions from "../utilities/functions";
+
+export interface BotOptions {
+	name: string;
+}
